hw06: refactored code
