+++ app.js
@@ -4,8 +4,8 @@ const cors = require("cors");
 
 require("dotenv").config();
 
-const authRouter = require("./routes/api/users");
 const contactsRouter = require("./routes/api/contacts");
+const authRouter = require("./routes/api/users");
 
 const app = express();
