+++ controllers/contacts/updateContact.js
@@ -1,0 +1,16 @@
+const { HttpError } = require("../../utils");
+const Contact = require("../../models/contact");
+
+const updateContact = async (req, res) => {
+    const { contactId } = req.params;
+    const contactUpDate = await Contact.findByIdAndUpdate(contactId, req.body, {
+        new: true,
+    });
+    console.log(contactUpDate);
+    if (!contactUpDate) {
+        throw HttpError(404, "Not found");
+    }
+    res.json(contactUpDate);
+};
+
+module.exports = updateContact;
