+++ controllers/users/getCurrentUser.js
@@ -1,0 +1,6 @@
+const getCurrentUser = async (req, res) => {
+    const { email, subscription } = req.user;
+    res.json({ email, subscription });
+};
+
+module.exports = getCurrentUser;
