+++ controllers/users/loginUser.test.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 const request = require("supertest");
-const app = require("../app");
+const app = require("../../app");
 
 const { DB_HOST } = process.env;
